+++ CHANGELOG.md
@@ -10,6 +10,8 @@
   - The Namespaces contract acts as a global register where process contracts register themselves
 - The Processes contract has been adapted to the new Namespaces contract
 - The `Processes` method `setResults` and `getResults` are now part of a future contract `Results`
+  - Oracles now set a process' results on the Results contract
+  - The results contract can then set the `RESULTS` status on the process contract
 - `EnsPublicResolver` is now renamed to `EnsResolver` in TypeScript
 - Validators now have a `bytes32` data type